validate template
10/11/2023

+++ trian/app/Http/Controllers/TemplateController.php
@@ -2,9 +2,14 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Event;
 use App\Models\Template;
+use App\Models\User;
+use Carbon\Carbon;
 use Illuminate\Http\Request;
-use Illuminate\Validation\Validator;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Validation\Rule;
+use Illuminate\Validation\ValidationException;
 
 class TemplateController extends Controller
 {
@@ -17,20 +22,55 @@ class TemplateController extends Controller
 
     public function create(Request $request)
     {
-
-        $data = $request->input('data');
+        // check null
+        $validator = Validator::make(
+            $request->all(),
+            [
+                'data' => 'required',
+                'event_id' => 'required|exists:events,id',
+                'timer' => 'required|date_format:Y-m-d H:i'
+            ],
+            [
+                'required' => 'Trường :attribute không được để trống.',
+                'email' => 'Trường :attribute phải là địa chỉ email hợp lệ.',
+                'unique' => 'Trường :attribute đã tồn tại.',
+                'exists' => 'Trường :attribute không hợp lệ.',
+                'date_format' => 'Trường :attribute không đúng định dạng Y-m-d'
+            ]
+        );
+        if ($validator->fails()) {
+            return response()->json(['error' => $validator->errors()], 422);
+        }
+        $list = $request->input('data');
         $event_id = $request->input('event_id');
         $timer = $request->input('timer'); // Không được lấy ngày giờ hiện tại nghen
-        // Tự động xác định giá trị cho trường "type" dựa trên dữ liệu
-        $dataArray = explode(',', $data);
-        if (count($dataArray) == 1) {
-            $type = 'USER';
-        } elseif (count($dataArray) >= 2) {
-            $type = 'GROUP';
-        } else {
-            return response()->json(['error' => 'Dữ liệu không hợp lệ.'], 400);
+        // bắt lỗi sự kiện
+        $currentTime = Carbon::now();
+        $inputTime = Carbon::createFromFormat('Y-m-d H:i', $timer);
+        if ($inputTime->diffInMinutes($currentTime) <= 30 || $inputTime->isPast()) {
+            return response()->json(['error' => 'Thời gian phải sau 30 phút so với thời gian hiện tại.'], 422);
         }
-
+        $event = Event::find($request->event_id);
+        if ($event->eventdate <= Carbon::now()) {
+            return response()->json(['error' => 'Sự kiện đã hoặc đang diễn.'], 422);
+        }
+        // Tự động xác định giá trị cho trường "type" dựa trên dữ liệu
+        $dataArray = explode(',', $list);
+        $users = User::WhereIn('id', $dataArray)->get();
+        $totalItems = count($users);
+        $counter = 0;
+        $data = "";
+        foreach ($users as $user) {
+            $data = $data . $user->id;
+            if (++$counter < $totalItems) {
+                $data = $data . ',';
+            }
+        }
+        if ($totalItems == 1) {
+            $type = 'USER';
+        } elseif ($totalItems >= 2) {
+            $type = 'GROUP';
+        } 
         // Tiến hành thêm dữ liệu
         $template = Template::create([
             'timer' => $timer,
@@ -38,7 +78,6 @@ class TemplateController extends Controller
             'data' => $data,
             'event_id' => $event_id,
         ]);
-
         return response()->json($template);
     }
 
@@ -46,31 +85,47 @@ class TemplateController extends Controller
     public function update($id, Request $request)
     {
         $template = Template::find($id);
-
-        if (!$template) {
-            return response()->json(['error' => 'Sự kiện không tồn tại.'], 404);
+        if (empty($template)) {
+            return response()->json(['error' => 'Template không tồn tại.'], 404);
         }
-
+        // check null
+        $validator = Validator::make(
+            $request->all(),
+            [
+                'data' => 'required',
+                'event_id' => 'required|exists:events,id',
+                'timer' => 'required|date_format:Y-m-d H:i'
+            ],
+            [
+                'required' => 'Trường :attribute không được để trống.',
+                'unique' => 'Trường :attribute đã tồn tại.',
+                'exists' => 'Trường :attribute không hợp lệ.',
+                'date_format' => 'Trường :attribute không đúng định dạng Y-m-d'
+            ]
+        );
+        if ($validator->fails()) {
+            return response()->json(['error' => $validator->errors()], 422);
+        }
         $timer = $request->input('timer'); // Không được lấy ngày giờ hiện tại nghen
         $data = $request->input('data');
-        $message_id = $request->input('message_id');
         $event_id = $request->input('event_id');
-
+        $currentTime = Carbon::now();
+        $inputTime = Carbon::createFromFormat('Y-m-d H:i', $timer);
+        if ($inputTime->diffInMinutes($currentTime) <= 30 || $inputTime->isPast()) {
+            return response()->json(['error' => 'Thời gian phải sau 30 phút so với thời gian hiện tại.'], 422);
+        }
         // Tự động xác định giá trị cho trường "type" dựa trên dữ liệu
         $dataArray = explode(',', $data);
         if (count($dataArray) == 1) {
             $type = 'USER';
         } elseif (count($dataArray) >= 2) {
             $type = 'GROUP';
-        } else {
-            return response()->json(['error' => 'Dữ liệu không hợp lệ.'], 400);
         }
 
         // Cập nhật dữ liệu cho sự kiện có ID tương ứng
         $template->timer = $timer;
         $template->type = $type;
         $template->data = $data;
-        $template->message_id = $message_id;
         $template->event_id = $event_id;
         $template->save();
 
@@ -78,16 +133,21 @@ class TemplateController extends Controller
     }
 
 
-    public function findId(Request $request)
-    {
-        $id = $request->input('id');
+    public function findId($id)
+    {    
         $template = Template::find($id);
+        if(empty($template)) {
+            return response()->json(['error'=> 'Template không tồn tại.'],404);
+        }
         return response()->json($template);
     }
 
     public function delete($id)
     {
         $template = Template::find($id);
+        if(empty($template)) {
+            return response()->json(['error'=> 'Template không tồn tại.'],404);
+        }
         try {
             // Xóa 
             $template->delete();
